Remove init file.
Move init functionality to DevModeHandler.
