+++ flow-server/src/main/java/com/vaadin/flow/server/VaadinServlet.java
@@ -30,7 +30,6 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.CurrentInstance;
 import com.vaadin.flow.server.ServletHelper.RequestType;
-import com.vaadin.flow.server.startup.DevModeInitializer;
 import com.vaadin.flow.server.webjar.WebJarServer;
 import com.vaadin.flow.shared.JsonConstants;
 
@@ -51,7 +50,7 @@ import com.vaadin.flow.shared.JsonConstants;
 public class VaadinServlet extends HttpServlet {
     private VaadinServletService servletService;
     private StaticFileHandler staticFileHandler;
-    private DevModeHandler devmodeHandler = DevModeInitializer.getDevModeHandler();
+    private DevModeHandler devmodeHandler = DevModeHandler.getDevModeHandler();
     private WebJarServer webJarServer;
 
     /**
