+++ flow-server/src/main/java/com/vaadin/flow/server/startup/ServletDeployer.java
@@ -33,6 +33,7 @@ import org.slf4j.LoggerFactory;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.server.DeploymentConfigurationFactory;
+import com.vaadin.flow.server.DevModeHandler;
 import com.vaadin.flow.server.VaadinServlet;
 import com.vaadin.flow.server.VaadinServletConfiguration;
 
@@ -120,7 +121,7 @@ public class ServletDeployer implements ServletContextListener {
                         "/frontend/*");
             }
         }
-        if(DevModeInitializer.getDevModeHandler() == null) {
+        if(DevModeHandler.getDevModeHandler() == null) {
             ServletRegistration vaadinServlet = findVaadinServlet(context);
 
             if(vaadinServlet != null) {
@@ -128,7 +129,7 @@ public class ServletDeployer implements ServletContextListener {
                         new StubServletConfig(context, vaadinServlet),
                         vaadinServlet.getClass());
 
-                DevModeInitializer.start(deploymentConfiguration);
+                DevModeHandler.start(deploymentConfiguration);
             }
         }
     }
